requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
sentence-transformers
faiss-cpu
pypdf
python-docx
sqlalchemy
databases
python-multipart